my pending assignment page created

+++ src/components/AssignmentCard/AssignmentCard.jsx
@@ -1,3 +1,4 @@
+import { useContext } from "react";
 import { BsBarChart } from "react-icons/bs";
 import { FaEdit } from "react-icons/fa";
 import { IoCheckmarkCircleOutline, IoEyeOutline } from "react-icons/io5";
@@ -5,42 +6,58 @@ import { MdOutlineDelete } from "react-icons/md";
 import { SlCalender } from "react-icons/sl";
 import { Link } from "react-router-dom";
 import Swal from "sweetalert2";
+import { authContext } from "../../providers/AuthProvider";
 
 
 const AssignmentCard = ({ assignment }) => {
 
     // console.log(assignment);
+    const { user } = useContext(authContext)
 
-    const { _id, title, marks, thumbnail, level, dueDate } = assignment
+    const { _id, title, marks, thumbnail, level, dueDate, email } = assignment
 
     const handleDelete = _id => {
-        // console.log(_id);
-        Swal.fire({
-            title: "Are you sure?",
-            text: "You won't be able to undo this!",
-            icon: "warning",
-            showCancelButton: true,
-            confirmButtonColor: "#3085d6",
-            cancelButtonColor: "#d33",
-            confirmButtonText: "Yes"
-        }).then((result) => {
-            if (result.isConfirmed) {
 
-                fetch(`http://localhost:5000/assignments/${_id}`, {
-                    method: 'DELETE'
-                })
-                    .then(res => res.json())
-                    .then(data => {
-                        if (data.deletedCount > 0) {
-                            Swal.fire({
-                                title: "Deleted!",
-                                text: "The Assignment has been deleted.",
-                                icon: "success"
-                            });
-                        }
+        // console.log(user.email);
+        // console.log(email);
+
+        if (user?.email == email) {
+            Swal.fire({
+                title: "Are you sure?",
+                text: "You won't be able to undo this!",
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#3085d6",
+                cancelButtonColor: "#d33",
+                confirmButtonText: "Yes"
+            }).then((result) => {
+                if (result.isConfirmed) {
+
+                    fetch(`http://localhost:5000/assignments/${_id}`, {
+                        method: 'DELETE'
                     })
-            }
-        });
+                        .then(res => res.json())
+                        .then(data => {
+                            if (data.deletedCount > 0) {
+                                Swal.fire({
+                                    title: "Deleted!",
+                                    text: "The Assignment has been deleted.",
+                                    icon: "success"
+                                });
+                            }
+                        })
+                }
+            });
+        }
+        else {
+            Swal.fire({
+                icon: "error",
+                title: "You can't remove this assignment!",
+                text: "You are only able to delete assignments that you've created.",
+                
+            });
+        }
+
     }
 
     return (
